requirements/dev.txt
Django==3.2.9
psycopg2-binary==2.9.1
djangorestframework==3.12.4
django-cors-headers==3.10.0
factory-boy==3.2.0
django-extensions==3.1.3
ipython==7.28.0
flake8==3.9.2
mypy==0.910
django-stubs==1.9.0
djangorestframework-stubs==1.4.0
tblib==1.7.0
bandit==1.7.0
coverage==6.0.1
requests==2.26.0
freezegun==1.1.0
types-requests==2.25.9
types-freezegun==1.1.0
drf-spectacular==0.20.1
django-filter==21.1
drf-nested-routers==0.93.4
